SP-488 refactor: 레포지토리 네이밍 컨벤션 맞추기
